requirements.txt
kafka-python==2.0.2
requests==2.25.1
psycopg2-binary==2.8.6
pyyaml==5.4.1
pytest==6.2.2
